# cmd/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A GEWIS Rooster maker",
        "title": "GRooster",
        "contact": {},
        "version": "0.1"
    },
    "paths": {
        "/auth/callback": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Validates state, exchanges code for token, and returns user info",
                "tags": [
                    "Auth"
                ],
                "summary": "Handle OAuth2 Callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "State returned from provider",
                        "name": "state",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization code from provider",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User info and token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request: missing or invalid state",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/redirect": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Generates state, sets a cookie, and redirects to Google OIDC",
                "tags": [
                    "Auth"
                ],
                "summary": "Redirect to OIDC provider",
                "parameters": [
                    {
                        "type": "string",
                        "description": "State returned from provider",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "redirect",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "pkg server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}